refactor Stream -> StreamCore/StreamExtra
